deps/v8/test/webkit/equality-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Test for equality of many combinations types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 0 == 0 is true
PASS 0 == 1 is false
PASS 0 == 0.1 is false
PASS 0 == 2 is false
PASS 0 == 3 is false
PASS 0 == 4 is false
PASS 0 == 5 is false
PASS 0 == 6 is false
PASS 0 == 7 is false
PASS 0 == -0 is true
PASS 0 == "0" is true
PASS 0 == "1" is false
PASS 0 == "0.1" is false
PASS 0 == "-0" is true
PASS 0 == null is false
PASS 0 == undefined is false
PASS 0 == false is true
PASS 0 == true is false
PASS 0 == new String("0") is true
PASS 0 == new Object is false
PASS 1 == 0 is false
PASS 1 == 1 is true
PASS 1 == 0.1 is false
PASS 1 == 2 is false
PASS 1 == 3 is false
PASS 1 == 4 is false
PASS 1 == 5 is false
PASS 1 == 6 is false
PASS 1 == 7 is false
PASS 1 == -0 is false
PASS 1 == "0" is false
PASS 1 == "1" is true
PASS 1 == "0.1" is false
PASS 1 == "-0" is false
PASS 1 == null is false
PASS 1 == undefined is false
PASS 1 == false is false
PASS 1 == true is true
PASS 1 == new String("0") is false
PASS 1 == new Object is false
PASS 0.1 == 0 is false
PASS 0.1 == 1 is false
PASS 0.1 == 0.1 is true
PASS 0.1 == 2 is false
PASS 0.1 == 3 is false
PASS 0.1 == 4 is false
PASS 0.1 == 5 is false
PASS 0.1 == 6 is false
PASS 0.1 == 7 is false
PASS 0.1 == -0 is false
PASS 0.1 == "0" is false
PASS 0.1 == "1" is false
PASS 0.1 == "0.1" is true
PASS 0.1 == "-0" is false
PASS 0.1 == null is false
PASS 0.1 == undefined is false
PASS 0.1 == false is false
PASS 0.1 == true is false
PASS 0.1 == new String("0") is false
PASS 0.1 == new Object is false
PASS 2 == 0 is false
PASS 2 == 1 is false
PASS 2 == 0.1 is false
PASS 2 == 2 is true
PASS 2 == 3 is false
PASS 2 == 4 is false
PASS 2 == 5 is false
PASS 2 == 6 is false
PASS 2 == 7 is false
PASS 2 == -0 is false
PASS 2 == "0" is false
PASS 2 == "1" is false
PASS 2 == "0.1" is false
PASS 2 == "-0" is false
PASS 2 == null is false
PASS 2 == undefined is false
PASS 2 == false is false
PASS 2 == true is false
PASS 2 == new String("0") is false
PASS 2 == new Object is false
PASS 3 == 0 is false
PASS 3 == 1 is false
PASS 3 == 0.1 is false
PASS 3 == 2 is false
PASS 3 == 3 is true
PASS 3 == 4 is false
PASS 3 == 5 is false
PASS 3 == 6 is false
PASS 3 == 7 is false
PASS 3 == -0 is false
PASS 3 == "0" is false
PASS 3 == "1" is false
PASS 3 == "0.1" is false
PASS 3 == "-0" is false
PASS 3 == null is false
PASS 3 == undefined is false
PASS 3 == false is false
PASS 3 == true is false
PASS 3 == new String("0") is false
PASS 3 == new Object is false
PASS 4 == 0 is false
PASS 4 == 1 is false
PASS 4 == 0.1 is false
PASS 4 == 2 is false
PASS 4 == 3 is false
PASS 4 == 4 is true
PASS 4 == 5 is false
PASS 4 == 6 is false
PASS 4 == 7 is false
PASS 4 == -0 is false
PASS 4 == "0" is false
PASS 4 == "1" is false
PASS 4 == "0.1" is false
PASS 4 == "-0" is false
PASS 4 == null is false
PASS 4 == undefined is false
PASS 4 == false is false
PASS 4 == true is false
PASS 4 == new String("0") is false
PASS 4 == new Object is false
PASS 5 == 0 is false
PASS 5 == 1 is false
PASS 5 == 0.1 is false
PASS 5 == 2 is false
PASS 5 == 3 is false
PASS 5 == 4 is false
PASS 5 == 5 is true
PASS 5 == 6 is false
PASS 5 == 7 is false
PASS 5 == -0 is false
PASS 5 == "0" is false
PASS 5 == "1" is false
PASS 5 == "0.1" is false
PASS 5 == "-0" is false
PASS 5 == null is false
PASS 5 == undefined is false
PASS 5 == false is false
PASS 5 == true is false
PASS 5 == new String("0") is false
PASS 5 == new Object is false
PASS 6 == 0 is false
PASS 6 == 1 is false
PASS 6 == 0.1 is false
PASS 6 == 2 is false
PASS 6 == 3 is false
PASS 6 == 4 is false
PASS 6 == 5 is false
PASS 6 == 6 is true
PASS 6 == 7 is false
PASS 6 == -0 is false
PASS 6 == "0" is false
PASS 6 == "1" is false
PASS 6 == "0.1" is false
PASS 6 == "-0" is false
PASS 6 == null is false
PASS 6 == undefined is false
PASS 6 == false is false
PASS 6 == true is false
PASS 6 == new String("0") is false
PASS 6 == new Object is false
PASS 7 == 0 is false
PASS 7 == 1 is false
PASS 7 == 0.1 is false
PASS 7 == 2 is false
PASS 7 == 3 is false
PASS 7 == 4 is false
PASS 7 == 5 is false
PASS 7 == 6 is false
PASS 7 == 7 is true
PASS 7 == -0 is false
PASS 7 == "0" is false
PASS 7 == "1" is false
PASS 7 == "0.1" is false
PASS 7 == "-0" is false
PASS 7 == null is false
PASS 7 == undefined is false
PASS 7 == false is false
PASS 7 == true is false
PASS 7 == new String("0") is false
PASS 7 == new Object is false
PASS -0 == 0 is true
PASS -0 == 1 is false
PASS -0 == 0.1 is false
PASS -0 == 2 is false
PASS -0 == 3 is false
PASS -0 == 4 is false
PASS -0 == 5 is false
PASS -0 == 6 is false
PASS -0 == 7 is false
PASS -0 == -0 is true
PASS -0 == "0" is true
PASS -0 == "1" is false
PASS -0 == "0.1" is false
PASS -0 == "-0" is true
PASS -0 == null is false
PASS -0 == undefined is false
PASS -0 == false is true
PASS -0 == true is false
PASS -0 == new String("0") is true
PASS -0 == new Object is false
PASS "0" == 0 is true
PASS "0" == 1 is false
PASS "0" == 0.1 is false
PASS "0" == 2 is false
PASS "0" == 3 is false
PASS "0" == 4 is false
PASS "0" == 5 is false
PASS "0" == 6 is false
PASS "0" == 7 is false
PASS "0" == -0 is true
PASS "0" == "0" is true
PASS "0" == "1" is false
PASS "0" == "0.1" is false
PASS "0" == "-0" is false
PASS "0" == null is false
PASS "0" == undefined is false
PASS "0" == false is true
PASS "0" == true is false
PASS "0" == new String("0") is true
PASS "0" == new Object is false
PASS "1" == 0 is false
PASS "1" == 1 is true
PASS "1" == 0.1 is false
PASS "1" == 2 is false
PASS "1" == 3 is false
PASS "1" == 4 is false
PASS "1" == 5 is false
PASS "1" == 6 is false
PASS "1" == 7 is false
PASS "1" == -0 is false
PASS "1" == "0" is false
PASS "1" == "1" is true
PASS "1" == "0.1" is false
PASS "1" == "-0" is false
PASS "1" == null is false
PASS "1" == undefined is false
PASS "1" == false is false
PASS "1" == true is true
PASS "1" == new String("0") is false
PASS "1" == new Object is false
PASS "0.1" == 0 is false
PASS "0.1" == 1 is false
PASS "0.1" == 0.1 is true
PASS "0.1" == 2 is false
PASS "0.1" == 3 is false
PASS "0.1" == 4 is false
PASS "0.1" == 5 is false
PASS "0.1" == 6 is false
PASS "0.1" == 7 is false
PASS "0.1" == -0 is false
PASS "0.1" == "0" is false
PASS "0.1" == "1" is false
PASS "0.1" == "0.1" is true
PASS "0.1" == "-0" is false
PASS "0.1" == null is false
PASS "0.1" == undefined is false
PASS "0.1" == false is false
PASS "0.1" == true is false
PASS "0.1" == new String("0") is false
PASS "0.1" == new Object is false
PASS "-0" == 0 is true
PASS "-0" == 1 is false
PASS "-0" == 0.1 is false
PASS "-0" == 2 is false
PASS "-0" == 3 is false
PASS "-0" == 4 is false
PASS "-0" == 5 is false
PASS "-0" == 6 is false
PASS "-0" == 7 is false
PASS "-0" == -0 is true
PASS "-0" == "0" is false
PASS "-0" == "1" is false
PASS "-0" == "0.1" is false
PASS "-0" == "-0" is true
PASS "-0" == null is false
PASS "-0" == undefined is false
PASS "-0" == false is true
PASS "-0" == true is false
PASS "-0" == new String("0") is false
PASS "-0" == new Object is false
PASS null == 0 is false
PASS null == 1 is false
PASS null == 0.1 is false
PASS null == 2 is false
PASS null == 3 is false
PASS null == 4 is false
PASS null == 5 is false
PASS null == 6 is false
PASS null == 7 is false
PASS null == -0 is false
PASS null == "0" is false
PASS null == "1" is false
PASS null == "0.1" is false
PASS null == "-0" is false
PASS null == null is true
PASS null == undefined is true
PASS null == false is false
PASS null == true is false
PASS null == new String("0") is false
PASS null == new Object is false
PASS undefined == 0 is false
PASS undefined == 1 is false
PASS undefined == 0.1 is false
PASS undefined == 2 is false
PASS undefined == 3 is false
PASS undefined == 4 is false
PASS undefined == 5 is false
PASS undefined == 6 is false
PASS undefined == 7 is false
PASS undefined == -0 is false
PASS undefined == "0" is false
PASS undefined == "1" is false
PASS undefined == "0.1" is false
PASS undefined == "-0" is false
PASS undefined == null is true
PASS undefined == undefined is true
PASS undefined == false is false
PASS undefined == true is false
PASS undefined == new String("0") is false
PASS undefined == new Object is false
PASS false == 0 is true
PASS false == 1 is false
PASS false == 0.1 is false
PASS false == 2 is false
PASS false == 3 is false
PASS false == 4 is false
PASS false == 5 is false
PASS false == 6 is false
PASS false == 7 is false
PASS false == -0 is true
PASS false == "0" is true
PASS false == "1" is false
PASS false == "0.1" is false
PASS false == "-0" is true
PASS false == null is false
PASS false == undefined is false
PASS false == false is true
PASS false == true is false
PASS false == new String("0") is true
PASS false == new Object is false
PASS true == 0 is false
PASS true == 1 is true
PASS true == 0.1 is false
PASS true == 2 is false
PASS true == 3 is false
PASS true == 4 is false
PASS true == 5 is false
PASS true == 6 is false
PASS true == 7 is false
PASS true == -0 is false
PASS true == "0" is false
PASS true == "1" is true
PASS true == "0.1" is false
PASS true == "-0" is false
PASS true == null is false
PASS true == undefined is false
PASS true == false is false
PASS true == true is true
PASS true == new String("0") is false
PASS true == new Object is false
PASS new String("0") == 0 is true
PASS new String("0") == 1 is false
PASS new String("0") == 0.1 is false
PASS new String("0") == 2 is false
PASS new String("0") == 3 is false
PASS new String("0") == 4 is false
PASS new String("0") == 5 is false
PASS new String("0") == 6 is false
PASS new String("0") == 7 is false
PASS new String("0") == -0 is true
PASS new String("0") == "0" is true
PASS new String("0") == "1" is false
PASS new String("0") == "0.1" is false
PASS new String("0") == "-0" is false
PASS new String("0") == null is false
PASS new String("0") == undefined is false
PASS new String("0") == false is true
PASS new String("0") == true is false
PASS new String("0") == new String("0") is false
PASS new String("0") == new Object is false
PASS new Object == 0 is false
PASS new Object == 1 is false
PASS new Object == 0.1 is false
PASS new Object == 2 is false
PASS new Object == 3 is false
PASS new Object == 4 is false
PASS new Object == 5 is false
PASS new Object == 6 is false
PASS new Object == 7 is false
PASS new Object == -0 is false
PASS new Object == "0" is false
PASS new Object == "1" is false
PASS new Object == "0.1" is false
PASS new Object == "-0" is false
PASS new Object == null is false
PASS new Object == undefined is false
PASS new Object == false is false
PASS new Object == true is false
PASS new Object == new String("0") is false
PASS new Object == new Object is false
PASS 0 === 0 is true
PASS 0 === 1 is false
PASS 0 === 0.1 is false
PASS 0 === 2 is false
PASS 0 === 3 is false
PASS 0 === 4 is false
PASS 0 === 5 is false
PASS 0 === 6 is false
PASS 0 === 7 is false
PASS 0 === -0 is true
PASS 0 === "0" is false
PASS 0 === "1" is false
PASS 0 === "0.1" is false
PASS 0 === "-0" is false
PASS 0 === null is false
PASS 0 === undefined is false
PASS 0 === false is false
PASS 0 === true is false
PASS 0 === new String("0") is false
PASS 0 === new Object is false
PASS 1 === 0 is false
PASS 1 === 1 is true
PASS 1 === 0.1 is false
PASS 1 === 2 is false
PASS 1 === 3 is false
PASS 1 === 4 is false
PASS 1 === 5 is false
PASS 1 === 6 is false
PASS 1 === 7 is false
PASS 1 === -0 is false
PASS 1 === "0" is false
PASS 1 === "1" is false
PASS 1 === "0.1" is false
PASS 1 === "-0" is false
PASS 1 === null is false
PASS 1 === undefined is false
PASS 1 === false is false
PASS 1 === true is false
PASS 1 === new String("0") is false
PASS 1 === new Object is false
PASS 0.1 === 0 is false
PASS 0.1 === 1 is false
PASS 0.1 === 0.1 is true
PASS 0.1 === 2 is false
PASS 0.1 === 3 is false
PASS 0.1 === 4 is false
PASS 0.1 === 5 is false
PASS 0.1 === 6 is false
PASS 0.1 === 7 is false
PASS 0.1 === -0 is false
PASS 0.1 === "0" is false
PASS 0.1 === "1" is false
PASS 0.1 === "0.1" is false
PASS 0.1 === "-0" is false
PASS 0.1 === null is false
PASS 0.1 === undefined is false
PASS 0.1 === false is false
PASS 0.1 === true is false
PASS 0.1 === new String("0") is false
PASS 0.1 === new Object is false
PASS 2 === 0 is false
PASS 2 === 1 is false
PASS 2 === 0.1 is false
PASS 2 === 2 is true
PASS 2 === 3 is false
PASS 2 === 4 is false
PASS 2 === 5 is false
PASS 2 === 6 is false
PASS 2 === 7 is false
PASS 2 === -0 is false
PASS 2 === "0" is false
PASS 2 === "1" is false
PASS 2 === "0.1" is false
PASS 2 === "-0" is false
PASS 2 === null is false
PASS 2 === undefined is false
PASS 2 === false is false
PASS 2 === true is false
PASS 2 === new String("0") is false
PASS 2 === new Object is false
PASS 3 === 0 is false
PASS 3 === 1 is false
PASS 3 === 0.1 is false
PASS 3 === 2 is false
PASS 3 === 3 is true
PASS 3 === 4 is false
PASS 3 === 5 is false
PASS 3 === 6 is false
PASS 3 === 7 is false
PASS 3 === -0 is false
PASS 3 === "0" is false
PASS 3 === "1" is false
PASS 3 === "0.1" is false
PASS 3 === "-0" is false
PASS 3 === null is false
PASS 3 === undefined is false
PASS 3 === false is false
PASS 3 === true is false
PASS 3 === new String("0") is false
PASS 3 === new Object is false
PASS 4 === 0 is false
PASS 4 === 1 is false
PASS 4 === 0.1 is false
PASS 4 === 2 is false
PASS 4 === 3 is false
PASS 4 === 4 is true
PASS 4 === 5 is false
PASS 4 === 6 is false
PASS 4 === 7 is false
PASS 4 === -0 is false
PASS 4 === "0" is false
PASS 4 === "1" is false
PASS 4 === "0.1" is false
PASS 4 === "-0" is false
PASS 4 === null is false
PASS 4 === undefined is false
PASS 4 === false is false
PASS 4 === true is false
PASS 4 === new String("0") is false
PASS 4 === new Object is false
PASS 5 === 0 is false
PASS 5 === 1 is false
PASS 5 === 0.1 is false
PASS 5 === 2 is false
PASS 5 === 3 is false
PASS 5 === 4 is false
PASS 5 === 5 is true
PASS 5 === 6 is false
PASS 5 === 7 is false
PASS 5 === -0 is false
PASS 5 === "0" is false
PASS 5 === "1" is false
PASS 5 === "0.1" is false
PASS 5 === "-0" is false
PASS 5 === null is false
PASS 5 === undefined is false
PASS 5 === false is false
PASS 5 === true is false
PASS 5 === new String("0") is false
PASS 5 === new Object is false
PASS 6 === 0 is false
PASS 6 === 1 is false
PASS 6 === 0.1 is false
PASS 6 === 2 is false
PASS 6 === 3 is false
PASS 6 === 4 is false
PASS 6 === 5 is false
PASS 6 === 6 is true
PASS 6 === 7 is false
PASS 6 === -0 is false
PASS 6 === "0" is false
PASS 6 === "1" is false
PASS 6 === "0.1" is false
PASS 6 === "-0" is false
PASS 6 === null is false
PASS 6 === undefined is false
PASS 6 === false is false
PASS 6 === true is false
PASS 6 === new String("0") is false
PASS 6 === new Object is false
PASS 7 === 0 is false
PASS 7 === 1 is false
PASS 7 === 0.1 is false
PASS 7 === 2 is false
PASS 7 === 3 is false
PASS 7 === 4 is false
PASS 7 === 5 is false
PASS 7 === 6 is false
PASS 7 === 7 is true
PASS 7 === -0 is false
PASS 7 === "0" is false
PASS 7 === "1" is false
PASS 7 === "0.1" is false
PASS 7 === "-0" is false
PASS 7 === null is false
PASS 7 === undefined is false
PASS 7 === false is false
PASS 7 === true is false
PASS 7 === new String("0") is false
PASS 7 === new Object is false
PASS -0 === 0 is true
PASS -0 === 1 is false
PASS -0 === 0.1 is false
PASS -0 === 2 is false
PASS -0 === 3 is false
PASS -0 === 4 is false
PASS -0 === 5 is false
PASS -0 === 6 is false
PASS -0 === 7 is false
PASS -0 === -0 is true
PASS -0 === "0" is false
PASS -0 === "1" is false
PASS -0 === "0.1" is false
PASS -0 === "-0" is false
PASS -0 === null is false
PASS -0 === undefined is false
PASS -0 === false is false
PASS -0 === true is false
PASS -0 === new String("0") is false
PASS -0 === new Object is false
PASS "0" === 0 is false
PASS "0" === 1 is false
PASS "0" === 0.1 is false
PASS "0" === 2 is false
PASS "0" === 3 is false
PASS "0" === 4 is false
PASS "0" === 5 is false
PASS "0" === 6 is false
PASS "0" === 7 is false
PASS "0" === -0 is false
PASS "0" === "0" is true
PASS "0" === "1" is false
PASS "0" === "0.1" is false
PASS "0" === "-0" is false
PASS "0" === null is false
PASS "0" === undefined is false
PASS "0" === false is false
PASS "0" === true is false
PASS "0" === new String("0") is false
PASS "0" === new Object is false
PASS "1" === 0 is false
PASS "1" === 1 is false
PASS "1" === 0.1 is false
PASS "1" === 2 is false
PASS "1" === 3 is false
PASS "1" === 4 is false
PASS "1" === 5 is false
PASS "1" === 6 is false
PASS "1" === 7 is false
PASS "1" === -0 is false
PASS "1" === "0" is false
PASS "1" === "1" is true
PASS "1" === "0.1" is false
PASS "1" === "-0" is false
PASS "1" === null is false
PASS "1" === undefined is false
PASS "1" === false is false
PASS "1" === true is false
PASS "1" === new String("0") is false
PASS "1" === new Object is false
PASS "0.1" === 0 is false
PASS "0.1" === 1 is false
PASS "0.1" === 0.1 is false
PASS "0.1" === 2 is false
PASS "0.1" === 3 is false
PASS "0.1" === 4 is false
PASS "0.1" === 5 is false
PASS "0.1" === 6 is false
PASS "0.1" === 7 is false
PASS "0.1" === -0 is false
PASS "0.1" === "0" is false
PASS "0.1" === "1" is false
PASS "0.1" === "0.1" is true
PASS "0.1" === "-0" is false
PASS "0.1" === null is false
PASS "0.1" === undefined is false
PASS "0.1" === false is false
PASS "0.1" === true is false
PASS "0.1" === new String("0") is false
PASS "0.1" === new Object is false
PASS "-0" === 0 is false
PASS "-0" === 1 is false
PASS "-0" === 0.1 is false
PASS "-0" === 2 is false
PASS "-0" === 3 is false
PASS "-0" === 4 is false
PASS "-0" === 5 is false
PASS "-0" === 6 is false
PASS "-0" === 7 is false
PASS "-0" === -0 is false
PASS "-0" === "0" is false
PASS "-0" === "1" is false
PASS "-0" === "0.1" is false
PASS "-0" === "-0" is true
PASS "-0" === null is false
PASS "-0" === undefined is false
PASS "-0" === false is false
PASS "-0" === true is false
PASS "-0" === new String("0") is false
PASS "-0" === new Object is false
PASS null === 0 is false
PASS null === 1 is false
PASS null === 0.1 is false
PASS null === 2 is false
PASS null === 3 is false
PASS null === 4 is false
PASS null === 5 is false
PASS null === 6 is false
PASS null === 7 is false
PASS null === -0 is false
PASS null === "0" is false
PASS null === "1" is false
PASS null === "0.1" is false
PASS null === "-0" is false
PASS null === null is true
PASS null === undefined is false
PASS null === false is false
PASS null === true is false
PASS null === new String("0") is false
PASS null === new Object is false
PASS undefined === 0 is false
PASS undefined === 1 is false
PASS undefined === 0.1 is false
PASS undefined === 2 is false
PASS undefined === 3 is false
PASS undefined === 4 is false
PASS undefined === 5 is false
PASS undefined === 6 is false
PASS undefined === 7 is false
PASS undefined === -0 is false
PASS undefined === "0" is false
PASS undefined === "1" is false
PASS undefined === "0.1" is false
PASS undefined === "-0" is false
PASS undefined === null is false
PASS undefined === undefined is true
PASS undefined === false is false
PASS undefined === true is false
PASS undefined === new String("0") is false
PASS undefined === new Object is false
PASS false === 0 is false
PASS false === 1 is false
PASS false === 0.1 is false
PASS false === 2 is false
PASS false === 3 is false
PASS false === 4 is false
PASS false === 5 is false
PASS false === 6 is false
PASS false === 7 is false
PASS false === -0 is false
PASS false === "0" is false
PASS false === "1" is false
PASS false === "0.1" is false
PASS false === "-0" is false
PASS false === null is false
PASS false === undefined is false
PASS false === false is true
PASS false === true is false
PASS false === new String("0") is false
PASS false === new Object is false
PASS true === 0 is false
PASS true === 1 is false
PASS true === 0.1 is false
PASS true === 2 is false
PASS true === 3 is false
PASS true === 4 is false
PASS true === 5 is false
PASS true === 6 is false
PASS true === 7 is false
PASS true === -0 is false
PASS true === "0" is false
PASS true === "1" is false
PASS true === "0.1" is false
PASS true === "-0" is false
PASS true === null is false
PASS true === undefined is false
PASS true === false is false
PASS true === true is true
PASS true === new String("0") is false
PASS true === new Object is false
PASS new String("0") === 0 is false
PASS new String("0") === 1 is false
PASS new String("0") === 0.1 is false
PASS new String("0") === 2 is false
PASS new String("0") === 3 is false
PASS new String("0") === 4 is false
PASS new String("0") === 5 is false
PASS new String("0") === 6 is false
PASS new String("0") === 7 is false
PASS new String("0") === -0 is false
PASS new String("0") === "0" is false
PASS new String("0") === "1" is false
PASS new String("0") === "0.1" is false
PASS new String("0") === "-0" is false
PASS new String("0") === null is false
PASS new String("0") === undefined is false
PASS new String("0") === false is false
PASS new String("0") === true is false
PASS new String("0") === new String("0") is false
PASS new String("0") === new Object is false
PASS new Object === 0 is false
PASS new Object === 1 is false
PASS new Object === 0.1 is false
PASS new Object === 2 is false
PASS new Object === 3 is false
PASS new Object === 4 is false
PASS new Object === 5 is false
PASS new Object === 6 is false
PASS new Object === 7 is false
PASS new Object === -0 is false
PASS new Object === "0" is false
PASS new Object === "1" is false
PASS new Object === "0.1" is false
PASS new Object === "-0" is false
PASS new Object === null is false
PASS new Object === undefined is false
PASS new Object === false is false
PASS new Object === true is false
PASS new Object === new String("0") is false
PASS new Object === new Object is false
PASS successfullyParsed is true

TEST COMPLETE
